makofiles: Use the poky scripts under MIT
The SDK scripts that were copied from poky are dual-licensed under GPL-2 or
MIT. In later versions, poky removed the GPL-2 option.

Using them under MIT is preferrable because the license can easily be
embedded in the files that are integrated in Elbe's build output.

Embed the MIT in the mako templates and add the SPDX license headers.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>